cv_review/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
pypdf>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=5.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
